+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.0
+
+- Will now install `rustup` if it does not exist in the environment.
+- Added musl support to `cargo-binstall`.
+
 # 0.6.0
 
 - Breaking: Cargo `bins` must provide the `cargo-` crate prefix manually. This change allows